chore(front): split big assistant.ts file
