+++ front/lib/agent_yaml_converter/schemas.ts
@@ -1,11 +1,9 @@
 import { z } from "zod";
 
 import { additionalConfigurationSchema } from "@app/components/agent_builder/AgentBuilderFormContext";
-import {
-  MODEL_IDS,
-  MODEL_PROVIDER_IDS,
-  REASONING_EFFORT_IDS,
-} from "@app/types/assistant/assistant";
+import { MODEL_IDS } from "@app/types/assistant/models/models";
+import { MODEL_PROVIDER_IDS } from "@app/types/assistant/models/providers";
+import { REASONING_EFFORT_IDS } from "@app/types/assistant/models/reasoning";
 
 export const agentYAMLBasicInfoSchema = z.object({
   handle: z.string().min(1, "Handle is required"),
