+++ front/types/api/internal/agent_configuration.ts
@@ -2,13 +2,11 @@ import * as t from "io-ts";
 import type { JSONSchema7 } from "json-schema";
 
 import { validateJsonSchema } from "@app/lib/utils/json_schemas";
-import type { SupportedModel } from "@app/types/assistant/assistant";
-import {
-  isSupportedModel,
-  ModelIdCodec,
-  ModelProviderIdCodec,
-  ReasoningEffortCodec,
-} from "@app/types/assistant/assistant";
+import { isSupportedModel } from "@app/types/assistant/assistant";
+import { ModelIdCodec } from "@app/types/assistant/models/models";
+import { ModelProviderIdCodec } from "@app/types/assistant/models/providers";
+import { ReasoningEffortCodec } from "@app/types/assistant/models/reasoning";
+import type { SupportedModel } from "@app/types/assistant/models/types";
 import { createRangeCodec } from "@app/types/shared/utils/iots_utils";
 import { TimeframeUnitCodec } from "@app/types/shared/utils/time_frame";
