+++ front/types/assistant/models/fireworks.ts
@@ -1,0 +1,43 @@
+import type { ModelConfigurationType } from "@app/types";
+
+export const FIREWORKS_DEEPSEEK_R1_MODEL_ID =
+  "accounts/fireworks/models/deepseek-r1" as const;
+export const FIREWORKS_KIMI_K2_INSTRUCT_MODEL_ID =
+  "accounts/fireworks/models/kimi-k2-instruct" as const;
+export const FIREWORKS_DEEPSEEK_R1_MODEL_CONFIG: ModelConfigurationType = {
+  providerId: "fireworks",
+  modelId: FIREWORKS_DEEPSEEK_R1_MODEL_ID,
+  displayName: "DeepSeek R1 (Fireworks)",
+  contextSize: 164_000,
+  recommendedTopK: 32,
+  recommendedExhaustiveTopK: 64,
+  largeModel: true,
+  description:
+    "DeepSeek's reasoning model (164k context, served via Fireworks).",
+  shortDescription: "DeepSeek R1 (reasoning model).",
+  isLegacy: false,
+  isLatest: false,
+  generationTokensCount: 2048,
+  supportsVision: false,
+  minimumReasoningEffort: "none",
+  maximumReasoningEffort: "none",
+  defaultReasoningEffort: "none",
+};
+export const FIREWORKS_KIMI_K2_INSTRUCT_MODEL_CONFIG: ModelConfigurationType = {
+  providerId: "fireworks",
+  modelId: FIREWORKS_KIMI_K2_INSTRUCT_MODEL_ID,
+  displayName: "Kimi K2 Instruct (Fireworks)",
+  contextSize: 131_072,
+  recommendedTopK: 32,
+  recommendedExhaustiveTopK: 64,
+  largeModel: true,
+  description: "Kimi's K2 Instruct model (131k context, served via Fireworks).",
+  shortDescription: "Kimi's K2 Instruct model.",
+  isLegacy: false,
+  isLatest: false,
+  generationTokensCount: 2048,
+  supportsVision: false,
+  minimumReasoningEffort: "light",
+  maximumReasoningEffort: "light",
+  defaultReasoningEffort: "light",
+};
