+++ front/types/index.ts
@@ -11,6 +11,23 @@ export * from "./assistant/avatar";
 export * from "./assistant/builder";
 export * from "./assistant/conversation";
 export * from "./assistant/generation";
+export * from "./assistant/models/anthropic";
+export * from "./assistant/models/deepseek";
+export * from "./assistant/models/embedding";
+export * from "./assistant/models/fireworks";
+export * from "./assistant/models/fireworks";
+export * from "./assistant/models/fireworks";
+export * from "./assistant/models/fireworks";
+export * from "./assistant/models/google_ai_studio";
+export * from "./assistant/models/mistral";
+export * from "./assistant/models/models";
+export * from "./assistant/models/noop";
+export * from "./assistant/models/openai";
+export * from "./assistant/models/providers";
+export * from "./assistant/models/reasoning";
+export * from "./assistant/models/togetherai";
+export * from "./assistant/models/types";
+export * from "./assistant/models/xai";
 export * from "./assistant/pubsub";
 export * from "./assistant/templates";
 export * from "./assistant/visualization";
@@ -68,9 +85,9 @@ export * from "./shared/utils/error_utils";
 export * from "./shared/utils/general";
 export * from "./shared/utils/global_error_handler";
 export * from "./shared/utils/hashing";
+export * from "./shared/utils/http_headers";
 export * from "./shared/utils/iots_utils";
 export * from "./shared/utils/string_utils";
-export * from "./shared/utils/http_headers";
 export * from "./shared/utils/structured_data";
 export * from "./shared/utils/time_frame";
 export * from "./shared/utils/url_utils";
